Safari-609.1.11.1/LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-disabled-fullscreen-and-pip-expected.txt

Testing MacOSInlineMediaControls with disabled fullscreen and picture-in-picture buttons.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.topLeftControlsBar.visible is false
PASS successfullyParsed is true

TEST COMPLETE
